requirements.txt
Django==5.2
pycryptodome==3.22.0
qrcode==8.1
Pillow==11.2.1
django-crispy-forms==2.4